requirements.txt
cairosvg>=2.5.2
natsort>=7.1.1
PyPDF2>=1.26.0
validators>=0.18.2
wget>=3.2
ffmpeg-python>=0.2.0